PMM-11957 Small refactor.
